Moving wizard and steps to a folder on continuous applications
