+++ app/forms/candidate_interface/continuous_applications/course_selection_wizard.rb
@@ -1,0 +1,13 @@
+module CandidateInterface
+  module ContinuousApplications
+    class CourseSelectionWizard < DfE::Wizard
+      steps do
+        [
+          { do_you_know_the_course: DoYouKnowTheCourseStep },
+          { go_to_find_explanation: GoToFindExplanationStep },
+          { provider_selection: ProviderSelectionStep },
+        ]
+      end
+    end
+  end
+end
