+++ app/forms/candidate_interface/continuous_applications/do_you_know_the_course_step.rb
@@ -1,0 +1,20 @@
+module CandidateInterface
+  module ContinuousApplications
+    class DoYouKnowTheCourseStep < DfE::WizardStep
+      attr_accessor :answer
+      validates :answer, presence: true
+
+      def self.permitted_params
+        [:answer]
+      end
+
+      def next_step
+        if answer == 'yes'
+          :provider_selection
+        elsif answer == 'no'
+          :go_to_find_explanation
+        end
+      end
+    end
+  end
+end
