+++ app/forms/candidate_interface/continuous_applications/go_to_find_explanation_step.rb
@@ -1,0 +1,6 @@
+module CandidateInterface
+  module ContinuousApplications
+    class GoToFindExplanationStep < DfE::WizardStep
+    end
+  end
+end
