+++ app/forms/candidate_interface/continuous_applications/provider_selection_step.rb
@@ -1,0 +1,6 @@
+module CandidateInterface
+  module ContinuousApplications
+    class ProviderSelectionStep < DfE::WizardStep
+    end
+  end
+end
